tests for PDB are working

+++ controllers/humiocluster_controller.go
@@ -39,7 +39,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/utils/strings/slices"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/go-logr/logr"
@@ -432,6 +431,18 @@ func (r *HumioClusterReconciler) hasNoUnusedNodePoolStatus(hc *humiov1alpha1.Hum
 		}
 	}
 	return true, 0
+}
+
+// validateClusterNodePools validates that we do not set both minAvailable and maxUnavailable for each node pool
+func (r *HumioClusterReconciler) validateClusterNodePools(hc *humiov1alpha1.HumioCluster) error {
+	for _, nodePool := range hc.Spec.NodePools {
+		if nodePool.PodDisruptionBudget != nil {
+			if nodePool.PodDisruptionBudget.MinAvailable != nil && nodePool.PodDisruptionBudget.MaxUnavailable != nil {
+				return fmt.Errorf("cannot set both minAvailable and maxUnavailable for node pool %s", nodePool.Name)
+			}
+		}
+	}
+	return nil
 }
 
 func (r *HumioClusterReconciler) ensureHumioClusterBootstrapToken(ctx context.Context, hc *humiov1alpha1.HumioCluster) (reconcile.Result, error) {
@@ -2396,19 +2407,78 @@ func (r *HumioClusterReconciler) podLabelsForHumio(name string) map[string]strin
 }
 
 // PDB implementation
+// shouldCreatePDBForNodePool determines if we should create a PDB for the node pool
+func shouldCreatePDBForNodePool(nodePool humiov1alpha1.HumioNodePoolSpec) bool {
+	// Only create a PDB if the node pool has a configured value for MinAvailable or MaxUnavailable
+	return nodePool.PodDisruptionBudget != nil && (nodePool.PodDisruptionBudget.MinAvailable != nil || nodePool.PodDisruptionBudget.MaxUnavailable != nil)
+}
+
+// pdbBelongsToCurrentNodePools checks if the given PDB labels match a node pool in the HumioCluster
+func (r *HumioClusterReconciler) pdbBelongsToCurrentNodePools(ctx context.Context, pdb policyv1.PodDisruptionBudget, hc *humiov1alpha1.HumioCluster) bool {
+	for _, np := range hc.Spec.NodePools {
+		expectedLabels := pdbSelectorLabels(hc.Name, np.Name)
+		if reflect.DeepEqual(pdb.Labels, expectedLabels) {
+			return true
+		}
+	}
+	r.Log.Info("PDB does not belong to current node pools", "pdbName", pdb.Name)
+	return false
+}
+
+// pdbSelectorLabels returns the standard label set for selecting pods within a node pool
+func pdbSelectorLabels(clusterName, nodePoolName string) map[string]string {
+	return map[string]string{
+		"app":        "humio",
+		"humio_cr":   clusterName,
+		"humio_pool": nodePoolName,
+		"managed-by": "humio-operator",
+	}
+}
+
+// cleanupOrphanedNodePoolPDBs deletes any PDB in the namespace that belongs to this cluster but
+// does not match any of the current node pool specs
+func (r *HumioClusterReconciler) cleanupOrphanedNodePoolPDBs(ctx context.Context, hc *humiov1alpha1.HumioCluster) error {
+
+	list := &policyv1.PodDisruptionBudgetList{}
+	// Use helper function LabelAppHumio to generate common label
+	err := r.List(ctx, list, client.InNamespace(hc.Namespace), client.MatchingLabels(kubernetes.LabelsForHumio(hc.Name))) // Filter by "managed-by" label
+	if err != nil {
+		r.Log.Error(err, "Failed to list PDBs for cleanup")
+		return fmt.Errorf("failed to list PDBs for cleanup: %w", err)
+	}
+
+	for _, pdb := range list.Items {
+		// If this PDB's labels do not match a current node pool, delete it
+		if !r.pdbBelongsToCurrentNodePools(ctx, pdb, hc) {
+			r.Log.Info("Deleting orphaned PDB", "pdb", pdb.Name, "namespace", pdb.Namespace)
+			if delErr := r.Delete(ctx, &pdb); delErr != nil && !k8serrors.IsNotFound(delErr) {
+				r.Log.Error(delErr, "Failed to delete orphaned PDB", "pdb", pdb.Name)
+				return fmt.Errorf("failed to delete orphaned PDB %s: %w", pdb.Name, delErr)
+			}
+			r.Log.Info("Successfully deleted orphaned PDB", "pdb", pdb.Name)
+		}
+	}
+	return nil
+}
+
 // reconcilePodDisruptionBudgets ensures the correct PodDisruptionBudgets for all node pools
 func (r *HumioClusterReconciler) reconcilePodDisruptionBudgets(ctx context.Context, hc *humiov1alpha1.HumioCluster) (ctrl.Result, error) {
-	logger := log.FromContext(ctx)
 
 	// (1) Create or update PDB for each node pool
 	for _, nodePool := range hc.Spec.NodePools {
+
+		if nodePool.PodDisruptionBudget == nil {
+			r.Log.Info("Skipping PDB reconciliation for node pool as PodDisruptionBudget is not configured", "nodePool", nodePool.Name)
+			continue
+		}
+
 		pdbName := fmt.Sprintf("%s-%s-pdb", hc.Name, nodePool.Name)
 
 		if !shouldCreatePDBForNodePool(nodePool) {
 			// If we shouldn't create, ensure it's deleted if it exists
 			existingPDB := &policyv1.PodDisruptionBudget{}
 			if err := r.Get(ctx, types.NamespacedName{Name: pdbName, Namespace: hc.Namespace}, existingPDB); err == nil {
-				logger.Info("Deleting PDB for node pool as it should not exist", "pdbName", pdbName, "nodePool", nodePool.Name)
+				r.Log.Info("Deleting PDB for node pool as it should not exist", "pdbName", pdbName, "nodePool", nodePool.Name)
 				if delErr := r.Delete(ctx, existingPDB); delErr != nil && !k8serrors.IsNotFound(delErr) {
 					return ctrl.Result{}, fmt.Errorf("unable to delete PDB %s: %w", pdbName, delErr)
 				}
@@ -2418,6 +2488,7 @@ func (r *HumioClusterReconciler) reconcilePodDisruptionBudgets(ctx context.Conte
 
 		desiredPDB, err := r.buildNodePoolPDB(ctx, hc, nodePool)
 		if err != nil {
+			r.Log.Error(err, "Failed to build PDB object", "pdbName", pdbName)
 			return ctrl.Result{}, err
 		}
 
@@ -2425,44 +2496,58 @@ func (r *HumioClusterReconciler) reconcilePodDisruptionBudgets(ctx context.Conte
 		err = r.Get(ctx, types.NamespacedName{Name: pdbName, Namespace: hc.Namespace}, existingPDB)
 		if err != nil {
 			if k8serrors.IsNotFound(err) {
-				logger.Info("Creating PDB for node pool", "pdbName", pdbName, "nodePool", nodePool.Name)
+				r.Log.Info("Creating PDB for node pool", "pdbName", pdbName, "nodePool", nodePool.Name)
 				if createErr := r.Create(ctx, desiredPDB); createErr != nil {
+					r.Log.Error(createErr, "Failed to create PDB", "pdbName", pdbName)
 					return ctrl.Result{}, fmt.Errorf("unable to create PDB %s: %w", pdbName, createErr)
 				}
+				r.Log.Info("Successfully created PDB", "pdbName", pdbName)
 			} else {
+				r.Log.Error(err, "Failed to get PDB", "pdbName", pdbName)
 				return ctrl.Result{}, fmt.Errorf("unable to get PDB %s: %w", pdbName, err)
 			}
 		} else {
 			// Update existing if needed
 			if !arePDBsEqual(existingPDB, desiredPDB) {
-				logger.Info("Updating PDB", "pdbName", pdbName)
+				r.Log.Info("Updating PDB", "pdbName", pdbName)
 				existingPDB.Spec = desiredPDB.Spec
 				// Labels don't need to be updated as they are part of metadata
 				if updateErr := r.Update(ctx, existingPDB); updateErr != nil {
+					r.Log.Error(updateErr, "Failed to update PDB", "pdbName", pdbName)
 					return ctrl.Result{}, fmt.Errorf("unable to update PDB %s: %w", pdbName, updateErr)
 				}
+				r.Log.Info("Successfully updated PDB", "pdbName", pdbName)
 			}
 		}
 	}
 
 	// (2) Clean up orphaned PDBs that no longer match any node pool
 	if err := r.cleanupOrphanedNodePoolPDBs(ctx, hc); err != nil {
+		r.Log.Error(err, "Failed to cleanup orphaned PDBs")
 		return ctrl.Result{}, err
 	}
 
+	r.Log.Info("Finished reconciling PDBs")
 	return ctrl.Result{}, nil
 }
 
 // buildNodePoolPDB constructs a PodDisruptionBudget for the specified node pool
 func (r *HumioClusterReconciler) buildNodePoolPDB(ctx context.Context, hc *humiov1alpha1.HumioCluster, nodePool humiov1alpha1.HumioNodePoolSpec) (*policyv1.PodDisruptionBudget, error) {
-	logger := log.FromContext(ctx)
-	pdbName := fmt.Sprintf("%s-%s-pdb", hc.Name, nodePool.Name)
+
+	// Construct PDB name, handling cluster-level PDB
+	var pdbName string
 	if nodePool.Name == "" {
-		pdbName = fmt.Sprintf("%s-pdb", hc.Name)
+		pdbName = fmt.Sprintf("%s-cluster-pdb", hc.Name) // Unique name for cluster-level PDB
+		r.Log.Info("Building cluster-level PDB", "pdbName", pdbName)
+	} else {
+		pdbName = fmt.Sprintf("%s-%s-pdb", hc.Name, nodePool.Name)
+		r.Log.Info("Building node pool PDB", "pdbName", pdbName, "nodePool", nodePool.Name)
 	}
 
-	// Use PDB-standard selector labels
+	// PDB selector labels, including the "managed-by" label
 	labels := pdbSelectorLabels(hc.Name, nodePool.Name)
+	labels["managed-by"] = "humio-operator"
+	r.Log.Info("PDB labels", "labels", labels)
 
 	pdb := &policyv1.PodDisruptionBudget{
 		ObjectMeta: metav1.ObjectMeta{
@@ -2478,84 +2563,49 @@ func (r *HumioClusterReconciler) buildNodePoolPDB(ctx context.Context, hc *humio
 	}
 
 	// Set only one of MinAvailable or MaxUnavailable based on NodePool configuration
-	if nodePool.MinAvailable != nil {
-		pdb.Spec.MinAvailable = nodePool.MinAvailable
-	} else if nodePool.MaxUnavailable != nil {
-		pdb.Spec.MaxUnavailable = nodePool.MaxUnavailable
+	if nodePool.PodDisruptionBudget.MinAvailable != nil {
+		pdb.Spec.MinAvailable = nodePool.PodDisruptionBudget.MinAvailable
+		r.Log.Info("Setting PDB minAvailable", "pdbName", pdbName, "minAvailable", pdb.Spec.MinAvailable)
+	} else if nodePool.PodDisruptionBudget.MaxUnavailable != nil {
+		pdb.Spec.MaxUnavailable = nodePool.PodDisruptionBudget.MaxUnavailable
+		r.Log.Info("Setting PDB maxUnavailable", "pdbName", pdbName, "maxUnavailable", pdb.Spec.MaxUnavailable)
 	} else {
-		// By default, if neither are specified and node pool has more than one node we set MaxUnavailable to 1.
-		if nodePool.NodeCount > 1 {
-			pdb.Spec.MaxUnavailable = &intstr.IntOrString{Type: intstr.Int, IntVal: 1}
-		} else {
-			logger.Info("Neither minAvailable nor maxUnavailable set and node pool has only one node, defaulting to maxUnavailable=0", "nodePool", nodePool.Name)
-			pdb.Spec.MaxUnavailable = &intstr.IntOrString{Type: intstr.Int, IntVal: 0}
-		}
+		// By default, if neither are specified we set MaxUnavailable to 1.
+		pdb.Spec.MaxUnavailable = &intstr.IntOrString{Type: intstr.Int, IntVal: 1}
+		r.Log.Info("Setting PDB maxUnavailable to default", "pdbName", pdbName, "maxUnavailable", pdb.Spec.MaxUnavailable)
 	}
+
+	// Set UnhealthyPodEvictionPolicy if configured
+	if nodePool.PodDisruptionBudget.UnhealthyPodEvictionPolicy != nil {
+		pdb.Spec.UnhealthyPodEvictionPolicy = (*policyv1.UnhealthyPodEvictionPolicyType)(nodePool.PodDisruptionBudget.UnhealthyPodEvictionPolicy)
+		r.Log.Info("Setting PDB unhealthyPodEvictionPolicy", "pdbName", pdbName, "unhealthyPodEvictionPolicy", *pdb.Spec.UnhealthyPodEvictionPolicy)
+	}
+
+	// Set controller reference for automatic garbage collection
+	if err := controllerutil.SetControllerReference(hc, pdb, r.Scheme()); err != nil {
+		r.Log.Error(err, "Failed to set controller reference for PDB", "pdbName", pdbName)
+		return nil, fmt.Errorf("failed to set controller reference for PDB %s: %w", pdbName, err)
+	}
+
+	r.Log.Info("Successfully built PDB object", "pdbName", pdbName)
 	return pdb, nil
-}
-
-// shouldCreatePDBForNodePool determines if we should create a PDB for the node pool
-func shouldCreatePDBForNodePool(nodePool humiov1alpha1.HumioNodePoolSpec) bool {
-	// Only create a PDB if the node pool has a configured value for MinAvailable or MaxUnavailable
-	return nodePool.MinAvailable != nil || nodePool.MaxUnavailable != nil
-}
-
-// cleanupOrphanedNodePoolPDBs deletes any PDB in the namespace that belongs to this cluster but
-// does not match any of the current node pool specs
-func (r *HumioClusterReconciler) cleanupOrphanedNodePoolPDBs(ctx context.Context, hc *humiov1alpha1.HumioCluster) error {
-	logger := log.FromContext(ctx)
-
-	list := &policyv1.PodDisruptionBudgetList{}
-	err := r.List(ctx, list, client.InNamespace(hc.Namespace), client.MatchingLabels{"app": "humio", "humio_cr": hc.Name})
-	if err != nil {
-		return fmt.Errorf("failed to list PDBs for cleanup: %w", err)
-	}
-
-	for _, pdb := range list.Items {
-		// If this PDB's labels do not match a current node pool, delete it
-		if !r.pdbBelongsToCurrentNodePools(ctx, pdb, hc) {
-			logger.Info("Deleting orphaned PDB", "pdb", pdb.Name)
-			if delErr := r.Delete(ctx, &pdb); delErr != nil && !k8serrors.IsNotFound(delErr) {
-				return fmt.Errorf("failed to delete orphaned PDB %s: %w", pdb.Name, delErr)
-			}
-		}
-	}
-	return nil
-}
-
-// pdbBelongsToCurrentNodePools checks if the given PDB labels match a node pool in the HumioCluster
-func (r *HumioClusterReconciler) pdbBelongsToCurrentNodePools(ctx context.Context, pdb policyv1.PodDisruptionBudget, hc *humiov1alpha1.HumioCluster) bool {
-	logger := log.FromContext(ctx)
-	for _, np := range hc.Spec.NodePools {
-		expectedLabels := pdbSelectorLabels(hc.Name, np.Name)
-		if reflect.DeepEqual(pdb.Labels, expectedLabels) {
-			return true
-		}
-	}
-	logger.Info("PDB does not belong to current node pools", "pdbName", pdb.Name)
-	return false
-}
-
-// pdbSelectorLabels returns the standard label set for selecting pods within a node pool
-func pdbSelectorLabels(clusterName, nodePoolName string) map[string]string {
-	return map[string]string{
-		"app":        "humio",
-		"humio_cr":   clusterName,
-		"humio_pool": nodePoolName,
-	}
 }
 
 // arePDBsEqual checks if two PDBs have the same spec, ignoring generated fields
 func arePDBsEqual(existing, desired *policyv1.PodDisruptionBudget) bool {
-	return reflect.DeepEqual(existing.Spec, desired.Spec)
-}
-
-// validateClusterNodePools validates that we do not set both minAvailable and maxUnavailable for each node pool
-func (r *HumioClusterReconciler) validateClusterNodePools(hc *humiov1alpha1.HumioCluster) error {
-	for _, nodePool := range hc.Spec.NodePools {
-		if nodePool.MinAvailable != nil && nodePool.MaxUnavailable != nil {
-			return fmt.Errorf("cannot set both minAvailable and maxUnavailable for node pool %s", nodePool.Name)
-		}
+	// Compare the specs. We don't care about metadata because the name is already compared
+	// and the labels are used to determine if the PDB should be updated.
+	if !reflect.DeepEqual(existing.Spec.MinAvailable, desired.Spec.MinAvailable) {
+		return false
 	}
-	return nil
+	if !reflect.DeepEqual(existing.Spec.MaxUnavailable, desired.Spec.MaxUnavailable) {
+		return false
+	}
+	if !reflect.DeepEqual(existing.Spec.Selector, desired.Spec.Selector) {
+		return false
+	}
+	if !reflect.DeepEqual(existing.Spec.UnhealthyPodEvictionPolicy, desired.Spec.UnhealthyPodEvictionPolicy) {
+		return false
+	}
+	return true
 }
